!refactor: split core.schemas -> all domains & shared kernel
