+++ src/api/auth.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter, Depends, Request
 from fastapi.security import OAuth2PasswordRequestForm
 
 from src.audit import AuditService, get_audit_service
+from src.audit.schemas import AuditAction, AuditResult
 from src.audit.service import extract_client_info
 from src.auth import fastapi_users
 from src.auth.backend import (
@@ -18,8 +19,7 @@ from src.auth.schemas import (
     UserRead,
     UserUpdate,
 )
-from src.core.schemas.audit import AuditAction, AuditResult
-from src.core.schemas.error import ErrorCode
+from src.shared.errors import ErrorCode
 from src.exceptions import BusinessException
 
 router = APIRouter()
