--- src/core/schemas/__init__.py
@@ -1,19 +1,0 @@
-from .mixins import TimestampMixin
-from .error import ErrorCode, error_code_to_http_status
-from .oauth import OAuthAccount
-from .registry import redis_keys
-from .rbac import Permission, Role, RolePermission, UserRole
-from .user import User
-
-__all__ = [
-    "ErrorCode",
-    "Permission",
-    "Role",
-    "RolePermission",
-    "UserRole",
-    "TimestampMixin",
-    "OAuthAccount",
-    "User",
-    "error_code_to_http_status",
-    "redis_keys",
-]
