--- src/core/schemas/user.py
@@ -1,15 +1,0 @@
-from sqlmodel import Field, SQLModel
-
-from .mixins import TimestampMixin
-
-
-class User(SQLModel, TimestampMixin, table=True):
-    __tablename__ = "users"
-
-    id: int | None = Field(default=None, primary_key=True)
-    username: str = Field(unique=True, index=True, max_length=50)
-    email: str = Field(unique=True, index=True, max_length=320)
-    hashed_password: str = Field(max_length=1024)
-    is_active: bool = Field(default=True)
-    is_superuser: bool = Field(default=False)
-    is_verified: bool = Field(default=False)
