+++ src/shared/errors.py
