+++ tests/integration/test_audit_auth.py
@@ -1,7 +1,7 @@
 from sqlalchemy import select
 
-from src.core.schemas.audit import AuditAction, AuditLog, AuditResult
-from src.core.schemas.error import ErrorCode
+from src.audit.schemas import AuditAction, AuditLog, AuditResult
+from src.shared.errors import ErrorCode
 
 
 async def test_login_success_creates_audit_log(test_client, test_user, test_db):
@@ -102,7 +102,7 @@ async def test_refresh_token_failure_creates_audit_log(test_client, test_db):
 async def test_refresh_token_inactive_user_creates_audit_log(
     test_client, test_user, test_db
 ):
-    from src.core.schemas import User
+    from src.auth.schemas import User
 
     login_response = await test_client.post(
         "/auth/jwt/login",
