+++ tests/integration/test_audit_rbac.py
@@ -4,10 +4,10 @@ from httpx import ASGITransport, AsyncClient
 from sqlalchemy import select
 
 from src.auth import require_permissions, require_roles
-from src.core.schemas import User
-from src.core.schemas.audit import AuditAction, AuditLog, AuditResult
-from src.core.schemas.error import ErrorCode
-from src.core.schemas.rbac import Permission, RolePermission, UserRole
+from src.auth.schemas import User
+from src.audit.schemas import AuditAction, AuditLog, AuditResult
+from src.shared.errors import ErrorCode
+from src.auth.schemas import Permission, RolePermission, UserRole
 
 
 @pytest.fixture
