+++ tests/integration/test_pagination.py
@@ -6,7 +6,7 @@ from httpx import ASGITransport, AsyncClient
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from src.core.schemas import User
+from src.auth.schemas import User
 from src.session import get_session
